Function swc_ecma_transforms_typescript::strip::strip

pub fn strip(top_level_mark: Mark) -> impl Fold + VisitMut

Strips type annotations out.

See the examples directory of the crate to see how you can transpile a typescript file to a javascript file.